requirements.txt
flask
fpdf
pillow
werkzeug